requirements.txt
cmake==3.20.5
contourpy==1.1.1
cycler==0.12.1
distro==1.9.0
fonttools==4.53.0
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.3.0+a115f4b
packaging==24.0
pillow==10.3.0
pipe==1.6.0
pqkmeans==1.0.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-build==0.13.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.4
zipp==3.19.1
